# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sport Api Swagger Documentation Anatation",
        "description": "Sport Api Documentation to build the frontend of this site",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yurayur1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sport.napaautomotive.uz/",
            "description": "Sport api server link"
        },
        {
            "url": "https://sport.napaautomotive.uz/",
            "description": "Sport api server link"
        }
    ],
    "paths": {
...
        "/api/pasport": {
            "get": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Get all data from Pasport database",
                "description": "Via this link All Pasport` datas come",
                "operationId": "pasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pasport"
                ],
                "summary": "Register a pasport",
                "operationId": "storePasport",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pnfl": {
                                        "description": "Type your pnfl",
                                        "type": "integer"
                                    },
                                    "pasport_seria": {
                                        "description": "Type your pasport seria",
                                        "type": "string"
                                    },
                                    "pasport_seria_code": {
                                        "description": "Type your pasport seria code",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/api/personal": {
            "get": {
                "tags": [
                    "Personal Info"
                ],
                "summary": "Get all data from Personal Info database",
                "description": "Via this link All Personal Infos` datas come",
                "operationId": "personal",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "tags": [
                    "Personal Info"
                ],
                "summary": "Register a Personal Info",
                "operationId": "storePersonal",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Type user id",
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "description": "Type your full name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Type your email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Type your phone",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Type your gender",
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "description": "Type the date of your birth",
                                        "type": "string"
                                    },
                                    "nationality": {
                                        "description": "Type the nationality",
                                        "type": "string"
                                    },
                                    "pasport_id": {
                                        "description": "Type the pasport id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/personal/{personalInfo}": {
            "get": {
                "tags": [
                    "Personal Info"
                ],
                "summary": "Get one data from personal info database",
                "description": "Via this link a personal`s data comes to show",
                "operationId": "showPersonal",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personalInfo",
                        "in": "path",
                        "description": "ID for personal data",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            }
        },
        "/api/personal/{personalInfo}/update": {
            "put": {
                "tags": [
                    "Personal Info"
                ],
                "summary": "Updated",
                "description": "Update this personal",
                "operationId": "updatePersonal",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personal",
                        "in": "path",
                        "description": "personal that to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "description": "Type user id",
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "description": "Type your full name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Type your email",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "Type your phone",
                                        "type": "string"
                                    },
                                    "gender": {
                                        "description": "Type your gender",
                                        "type": "string"
                                    },
                                    "birth_date": {
                                        "description": "Type the date of your birth",
                                        "type": "string"
                                    },
                                    "nationality": {
                                        "description": "Type the nationality",
                                        "type": "string"
                                    },
                                    "pasport_id": {
                                        "description": "Type the pasport id",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid personal supplied"
                    },
                    "404": {
                        "description": "personal not found"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/personal/{personalInfo}/delete": {
            "delete": {
                "tags": [
                    "Personal Info"
                ],
                "summary": "Deletes a personal",
                "operationId": "destroyPersonal",
                "parameters": [
                    {
                        "name": "Accept-Language",
                        "in": "header",
                        "description": "Set language parameter by typing uz, ru, en",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "personalInfo",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid personal supplied"
                    },
                    "404": {
                        "description": "personal not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}